Travis CI uses working directory for temporary test files
Hook tests were failing in Travis CI when trying
to execute files written to /tmp with OS error 26
(text file busy).  I suspect this has something to do
with how the /tmp directory is configured in Travis,
so use the build directory instead.

The TMPDIR environment variable controls where
the Rust `tempfile` crate uses as the root temporary
directory, so we can set that globally to the Travis
build directory.