Do not overwrite existing hook
If for some reason the hook file already exists,
it would be overwritten with the hook template.

This could happen if, for example, the user had
copied a hook directory from another repo
before initializing the devlog repository.

This commit skips creation of the hook if
the file already exists to avoid overwriting
the user's changes.